feat(lang.converter.from_fhy_source): Extract out error listener from unit test conftest [ACTCOMP-41]
