+++ src/fhy/lang/converter/from_fhy_source.py
@@ -31,12 +31,18 @@
 
 """FhY source code to AST module converter."""
 
-from antlr4 import (  # type: ignore[import-untyped]
-    BailErrorStrategy,
+import logging
+from typing import List, Set
+
+from antlr4 import (  # type: ignore[import-untyped]  # type: ignore[import-untyped]
+    DFA,
     CommonTokenStream,
     InputStream,
 )
-from antlr4.error.ErrorListener import ErrorListener  # type: ignore[import-untyped]
+from antlr4.atn.ATNConfigSet import ATNConfigSet  # type: ignore[import-untyped]
+from antlr4.error.ErrorListener import (  # type: ignore[import-untyped]
+    ErrorListener,
+)
 
 from fhy import error
 from fhy.lang import ast
@@ -44,49 +50,188 @@ from fhy.lang.parser import FhYLexer, FhYParser
 
 from .from_parse_tree import from_parse_tree
 
+_log = logging.getLogger(__name__)
+
 
 class ThrowingErrorListener(ErrorListener):
-    """Custom Antlr error listener."""
+    """An Overly Verbose, Descriptive Antlr Error Listener for Reasons."""
 
-    def syntaxError(self, recognizer, offendingSymbol, line, column, msg, e):
-        message = f"Syntax error at {line}:{column} - {msg}"
+    log: logging.Logger
+
+    def __init__(self, log: logging.Logger = _log) -> None:
+        super().__init__()
+        self.log = log
+
+    def syntaxError(
+        self,
+        recognizer: FhYParser,
+        offendingSymbol: str,
+        line: int,
+        column: int,
+        msg: str,
+        e: Exception,
+    ):
+        text = self.get_text(recognizer, None, None)
+        context = type(recognizer._ctx).__name__
+        message = f'context={context}(Line {line}:{column}) input="{offendingSymbol}" '
+        # result = ParseTreeConverter().visit(recognizer._ctx)
+        # message += f'text="{text}" - msg={msg} - ast={result}'
+        message += f'text="{text}" - msg={msg}'
+
+        self.log.error(message)
 
         raise error.FhYSyntaxError(message) from e
 
+    def reportAmbiguity(
+        self,
+        recognizer: FhYParser,
+        dfa: DFA,
+        startIndex: int,
+        stopIndex: int,
+        exact: bool,
+        ambigAlts: Set[int],
+        configs: ATNConfigSet,
+    ):
+        report = self._report(
+            recognizer,
+            dfa,
+            startIndex,
+            stopIndex,
+            configs,
+        )
+        report += f" exact={exact}"
+        self.log.debug(report)
 
-# TODO: Extract out Construction of CST to another module within fhy/lang...
-def create_lexer(input_str: str) -> FhYLexer:
-    """Constructs the FhyLexer from input string source code."""
+    def reportAttemptingFullContext(
+        self,
+        recognizer: FhYParser,
+        dfa: DFA,
+        startIndex: int,
+        stopIndex: int,
+        conflictingAlts: Set[int],
+        configs: ATNConfigSet,
+    ):
+        report = self._report(
+            recognizer,
+            dfa,
+            startIndex,
+            stopIndex,
+            configs,
+        )
+        self.log.debug(report)
+
+    def reportContextSensitivity(
+        self,
+        recognizer: FhYParser,
+        dfa: DFA,
+        startIndex: int,
+        stopIndex: int,
+        prediction: int,
+        configs: ATNConfigSet,
+    ):
+        msg = self._report(
+            recognizer,
+            dfa,
+            startIndex,
+            stopIndex,
+            configs,
+        )
+        msg += f" predict={prediction}"
+        self.log.debug(msg)
+
+    def _report(
+        self,
+        recognizer: FhYParser,
+        dfa: DFA,
+        startIndex: int,
+        stopIndex: int,
+        configs: ATNConfigSet,
+    ) -> str:
+        decision = self.get_decision(recognizer, dfa)
+        conflict = self.get_conflict(recognizer, configs)
+        text = self.get_text(recognizer, startIndex, stopIndex)
+        if (ctx := recognizer._ctx) is None:
+            context = ""
+        else:
+            context = ctx.getText()
+
+        message = (
+            f"context={context}, d={decision}: ambigAlts={conflict}, input='{text}'"
+        )
+
+        return message
+
+    def get_text(self, recognizer: FhYParser, start: int, stop: int) -> str:
+        stream = recognizer.getTokenStream()
+        text = stream.getText(start, stop)
+
+        return text
+
+    def get_decision(self, recognizer: FhYParser, dfa: DFA) -> str:
+        decision: int = dfa.decision
+        index: int = dfa.atnStartState.ruleIndex
+
+        rules: List[str] = recognizer.ruleNames
+        if index < 0 or index >= len(rules):
+            return str(decision)
+
+        name = rules[index]
+        if not name:
+            return str(decision)
+
+        return name
+
+    def get_conflict(self, reportedAlts: Set[int], configs: ATNConfigSet) -> Set[int]:
+        if reportedAlts is None:
+            return reportedAlts
+
+        return {i.alt for i in configs}
+
+
+def create_lexer(input_str: str, log: logging.Logger = _log) -> FhYLexer:
+    """Construct the FhyLexer from input string source code."""
     input_stream = InputStream(input_str)
     lexer = FhYLexer(input_stream)
     lexer.removeErrorListeners()
-    lexer.addErrorListener(ThrowingErrorListener())
+    lexer.addErrorListener(ThrowingErrorListener(log))
 
     return lexer
 
 
-def create_parser(input_str: str) -> FhYParser:
-    """Constructs the FhyParser from input string source code."""
-    lexer = create_lexer(input_str)
+def create_parser(input_str: str, log: logging.Logger = _log) -> FhYParser:
+    """Construct the FhyParser from input string source code."""
+    lexer = create_lexer(input_str, log)
     token_stream = CommonTokenStream(lexer)
     parser = FhYParser(token_stream)
-    parser._errHandler = BailErrorStrategy()
+    # parser._errHandler = BailErrorStrategy()
     parser.removeErrorListeners()
-    parser.addErrorListener(ThrowingErrorListener())
+    parser.addErrorListener(ThrowingErrorListener(log))
 
     return parser
 
 
-def _fhy_source_to_parse_tree(fhy_source_content: str) -> FhYParser.ModuleContext:
-    fhy_parser = create_parser(fhy_source_content)
+def _fhy_source_to_parse_tree(
+    fhy_source_content: str, log: logging.Logger = _log
+) -> FhYParser.ModuleContext:
+    fhy_parser = create_parser(fhy_source_content, log)
     tree = fhy_parser.module()
 
     return tree
 
 
-def from_fhy_source(fhy_source_content: str) -> ast.Module:
-    """Convert FhY source code to an AST module."""
-    tree = _fhy_source_to_parse_tree(fhy_source_content)
+def from_fhy_source(fhy_source_content: str, log: logging.Logger = _log) -> ast.Module:
+    """Convert FhY source code into corresponding AST module representation.
+
+    Args:
+        fhy_source_content (str): FhY source code text.
+        log (logging.Logger): Inject a logger to control debugging information during
+            parsing.
+
+    Returns:
+        (ast.Module): AST module representation of input source code.
+
+    """
+    tree = _fhy_source_to_parse_tree(fhy_source_content, log)
     _ast = from_parse_tree(tree)
 
     return _ast
